r=int(input()) area=(3.14)*(r**2) print(area)

+++ Area_of_a_circle.py
@@ -1,3 +1,3 @@
 r=int(input())
-area=float((3.14) * r**2)
+area=(3.14)*(r**2)
 print(area)
